Update README.md
Initial contribution describing the project goal and objectives. I will need to include a link to the Notion page I created for this project and add another for the Flux.AI project repo. Working out of a phone currently so it will be difficult to do so right now but will make an attempt